requirements.txt
gradio==3.27.0
numpy==1.23.5
pandas==1.5.3
pandas_stubs==1.2.0.35
Pillow==9.5.0
sentence_transformers==2.2.2
pyarrow
transformers~=4.26.1
tqdm
streamlit
s3fs
requests
pinecone-client
